Extract paint shaders to files.
